refactor: extract helpers from sendChat and fix missing returns
- Rename throwError to displayError (it doesn't actually throw)
- Extract utility functions to index_util.svelte.ts
- Add missing return false after displayError calls (bug fix)
- Reorganize group/single chat branching logic with comments
- Add stage section comments for better readability
